+++ src/ronto/cli/build.py
@@ -11,7 +11,7 @@ from ronto.model.docker import docker_context
 
 @docker_context()
 def process(args):
-    verbose('Process build command')
+    verbose("Process build command")
     if args.interactive:
         builder = InteractiveBuilder(get_model())
     else:
@@ -20,18 +20,30 @@ def process(args):
 
 
 def add_command(subparser):
-    parser_pin = subparser.add_parser('build', help="""
+    parser_pin = subparser.add_parser(
+        "build",
+        help="""
             Actually build something
 
             source the environment and use bitbake.
-            """)
-    parser_pin.add_argument('-i', '--interactive',
-            help='Iignore targets and operate interactively',
-            action="store_true")
-    parser_pin.add_argument('-t', '--targets',
-            help='File that contains target definitions',
-            action="store_true")
-    parser_pin.add_argument('-b', '--clean-build',
-            help='Remove build directory directory',
-            action="store_true")
+            """,
+    )
+    parser_pin.add_argument(
+        "-i",
+        "--interactive",
+        help="Iignore targets and operate interactively",
+        action="store_true",
+    )
+    parser_pin.add_argument(
+        "-t",
+        "--targets",
+        help="File that contains target definitions",
+        action="store_true",
+    )
+    parser_pin.add_argument(
+        "-b",
+        "--clean-build",
+        help="Remove build directory directory",
+        action="store_true",
+    )
     parser_pin.set_defaults(func=process)