+++ CHANGELOG.rst
@@ -8,3 +8,6 @@ Version History
 
 0.0.1
     * Project created.
+0.1.0
+    * Rontofile version 1 (not checked)
+    * init, build and arbitrary docker sub-commands